+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.10.8] - 2020-07-29
+### Changed
+- if getUserModelByRoute function set, it determines the model for user in logs crud
+- changed usercrud abort route to method for artisan optimize
+
 ## [0.10.7] - 2020-07-21
 ### Updated
 - localizeArray -> added namespace option for lang keys